KEYCLOAK-8044 Clear theme caches on hot-deploy
